* Upgraded our included wxWidgets to 2.6.1, and modified some engine     code to use new wxWidgets APIs (bug #1712, partial).  We also     reapplied the following patches which had been backed out before     the upgrade:       - wx.rc: Build dependency hack.       - stc: Lots of local LISP-mode tweaking.       - evtloop: Hacked in *another* set of event loop hooks.
git-svn-id: https://imlsrc.dartmouth.edu/svn/public/halyard/trunk@28339 b24aca02-f418-0410-9c49-835e4aa915c3

+++ libs/wxIE/IEHtmlWin.cpp
@@ -117,7 +117,7 @@ public:
 	{
 		if (prepend.size() > 0)
 		{
-			int n = min(prepend.size(), cb);
+			int n = min(prepend.size(), static_cast<size_t>(cb));
 			prepend.copy((char *) pv, n);
 			prepend = prepend.substr(n);
 			if (pcbRead)
